#1764 exstracted filetering logic from view to util, it will be reused by PDF
