+++ view/statistics-flow-certificates.js
@@ -20,6 +20,7 @@ var _                 = require('mano').i18n.bind('View: Statistics')
   , pageQuery         = require('../utils/query/date-constrained-page')
   , copyDbDate        = require('../utils/copy-db-date')
   , queryServer       = require('./utils/statistics-flow-query-server')
+  , filterData        = require('../utils/statistics-flow-certificates-filter-result')
   , incrementDateByTimeUnit = require('../utils/increment-date-by-time-unit')
   , floorToTimeUnit         = require('../utils/floor-to-time-unit')
   , calculateDurationByMode = require('../utils/calculate-duration-by-mode');
@@ -48,59 +49,6 @@ db.BusinessProcess.extensions.forEach(function (ServiceType) {
 	});
 });
 
-var buildResultRow = function (rowData) {
-	return assign({
-		submitted: 0,
-		pending: 0,
-		pickup: 0,
-		withdrawn: 0,
-		rejected: 0,
-		sentBack: 0
-	}, rowData || {});
-};
-
-var accumulateResultRows = function (rows) {
-	var result = buildResultRow(rows[0]);
-	rows.slice(1).forEach(function (row) {
-		Object.keys(row).forEach(function (propertyName) {
-			result[propertyName] += row[propertyName];
-		});
-	});
-
-	return result;
-};
-
-var buildFilteredResult = function (data, key, service, certificate) {
-	if (!data[key]) return buildResultRow();
-	if (service) {
-		if (!data[key][service]) return buildResultRow();
-		if (certificate) {
-			return buildResultRow(data[key][service].certificate[certificate]);
-		}
-		return buildResultRow(data[key][service].businessProcess);
-	}
-	var rowsToAccumulate = [];
-	Object.keys(data[key]).forEach(function (service) {
-		if (certificate) {
-			if (!data[key][service].certificate[certificate]) return;
-			rowsToAccumulate.push(data[key][service].certificate[certificate]);
-		} else {
-			rowsToAccumulate.push(data[key][service].businessProcess);
-		}
-	});
-	return accumulateResultRows(rowsToAccumulate);
-};
-
-var filterData = function (data, query) {
-	var result = {};
-
-	Object.keys(data).forEach(function (date) {
-		result[date] = buildFilteredResult(data, date, query.service, query.certificate);
-	});
-
-	return result;
-};
-
 exports['statistics-main'] = function () {
 	var queryHandler, data = new ObservableValue({})
 	  , pagination = new Pagination('/flow/'), handlerConf;
@@ -115,7 +63,7 @@ exports['statistics-main'] = function () {
 		  , currentDate, offset, timeUnitsCount = 0, durationInTimeUnits, page;
 
 		dateFrom = query.dateFrom;
-		dateTo   = query.dateTo || new db.Date();
+		dateTo   = query.dateTo;
 		mode     = query.mode;
 		page     = query.page;
